+++ classes/mumie_problem.php
@@ -99,7 +99,7 @@ class mumie_problem implements \JsonSerializable {
      * @return mumie_problem
      */
     public static function from_task_db_object($data) {
-        $task = new \stdClass;
+        $task = new \stdClass();
         $task->link = \mod_mumie\locallib::remove_params_from_url($data->taskurl);
         $headline = [((object)["language" => $data->language, "name" => $data->name])];
         $task->headline = $headline;